move createDocketSwitchRulingTask to docketSwitchSlice.js; add completeDocketSwitchMailTask to docketSwitchSlice.js;
