+++ client/app/queue/docketSwitch/docketSwitchSlice.js
@@ -1,6 +1,7 @@
 import { createSlice, createAsyncThunk } from '@reduxjs/toolkit';
 
 import ApiUtil from '../../util/ApiUtil';
+import { onReceiveAmaTasks } from '../QueueActions';
 
 const initialState = {
   step: 0,
@@ -50,6 +51,40 @@ export const completeDocketSwitchGranted = createAsyncThunk(
       return result;
     } catch (error) {
       console.error('Error granting docket switch', error);
+      throw error;
+    }
+  }
+);
+
+export const createDocketSwitchRulingTask = createAsyncThunk(
+  'tasks/createDocketSwitchRulingTask',
+  async (data) => {
+    try {
+      const res = await ApiUtil.post('/tasks', { data });
+      const updatedTasks = res.body?.tasks?.data;
+
+      return updatedTasks;
+    } catch (error) {
+      console.error('Error creating task', error);
+      throw error;
+    }
+  }
+);
+
+export const completeDocketSwitchMailTask = createAsyncThunk(
+  'tasks/completeDocketSwitchRecommendationTask',
+  async ({ taskId }, { dispatch }) => {
+    try {
+      const res = await ApiUtil.patch(`/tasks/${taskId}`, {
+        data: { task: { status: 'completed' } },
+      });
+      const updatedTasks = res.body?.tasks?.data;
+
+      dispatch(onReceiveAmaTasks(updatedTasks));
+
+      return updatedTasks;
+    } catch (error) {
+      console.error('Error creating task', error);
       throw error;
     }
   }
